Feat: Add Search/lookup function and remove insert.go
