+++ cmd/radFS/arttest/main.go
@@ -9,7 +9,15 @@ func main() {
 
 	t.Insert([]byte("cat"), "v1")
 	t.Insert([]byte("car"), "v2")
-	t.Insert([]byte("cap"), "v3")
+	t.Insert([]byte("cab"), "v3")
+	v, ok := t.Search([]byte("cat"))
+	println("cat:", v, ok)
+
+	v, ok = t.Search([]byte("cab"))
+	println("cab:", v, ok)
+
+	v, ok = t.Search([]byte("cart"))
+	println("cart:", v, ok)
 
 	art.PrintTree(t.Root(), 0)
 }
